chore: resolve lint warnings in game UI
