+++ src/components/game/hud/EdictsPanel.tsx
